pyproject.toml
[project]
name = "autismbench"
version = "1.0.0"
description = "Constraint satisfaction benchmark for LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "plotly>=6.0.0",
    "numpy>=1.24.0",
    "nltk>=3.8.0",
]
